Move input utilities from native to Common/Input.
