+++ Qt/QtMain.h
@@ -27,8 +27,8 @@ QTM_USE_NAMESPACE
 #include "file/zip_read.h"
 #include "gfx/gl_common.h"
 #include "gfx_es2/gpu_features.h"
-#include "input/input_state.h"
-#include "input/keycodes.h"
+#include "Common/Input/InputState.h"
+#include "Common/Input/KeyCodes.h"
 #include "thin3d/thin3d.h"
 #include "base/NativeApp.h"
 #include "net/resolve.h"
