+++ SDL/SDLMain.cpp
@@ -24,8 +24,8 @@ SDLJoystick *joystick = NULL;
 #include "Common/TimeUtil.h"
 #include "ext/glslang/glslang/Public/ShaderLang.h"
 #include "image/png_load.h"
-#include "input/input_state.h"
-#include "input/keycodes.h"
+#include "Common/Input/InputState.h"
+#include "Common/Input/KeyCodes.h"
 #include "net/resolve.h"
 #include "NKCodeFromSDL.h"
 #include "util/const_map.h"
