+++ UI/ControlMappingScreen.cpp
@@ -23,8 +23,8 @@
 #include "base/display.h"
 #include "gfx/texture_atlas.h"
 #include "i18n/i18n.h"
-#include "input/keycodes.h"
-#include "input/input_state.h"
+#include "Common/Input/KeyCodes.h"
+#include "Common/Input/InputState.h"
 #include "ui/root.h"
 #include "ui/ui.h"
 #include "ui/ui_context.h"
