+++ UWP/CommonUWP/CommonUWP.vcxproj
@@ -386,6 +386,9 @@
     <ClInclude Include="..\..\Common\BitScan.h" />
     <ClInclude Include="..\..\Common\BitSet.h" />
     <ClInclude Include="..\..\Common\Buffer.h" />
+    <ClInclude Include="..\..\Common\Input\GestureDetector.h" />
+    <ClInclude Include="..\..\Common\Input\InputState.h" />
+    <ClInclude Include="..\..\Common\Input\KeyCodes.h" />
     <ClInclude Include="..\..\Common\Serialize\Serializer.h" />
     <ClInclude Include="..\..\Common\Serialize\SerializeDeque.h" />
     <ClInclude Include="..\..\Common\Serialize\SerializeFuncs.h" />
@@ -422,6 +425,10 @@
     <ClInclude Include="..\..\Common\SysError.h" />
     <ClInclude Include="..\..\Common\ThreadPools.h" />
     <ClInclude Include="..\..\Common\ThreadSafeList.h" />
+    <ClInclude Include="..\..\Common\Thread\Executor.h" />
+    <ClInclude Include="..\..\Common\Thread\PrioritizedWorkQueue.h" />
+    <ClInclude Include="..\..\Common\Thread\ThreadPool.h" />
+    <ClInclude Include="..\..\Common\Thread\ThreadUtil.h" />
     <ClInclude Include="..\..\Common\Thunk.h" />
     <ClInclude Include="..\..\Common\Timer.h" />
     <ClInclude Include="..\..\Common\TimeUtil.h" />
@@ -442,6 +449,8 @@
     <ClCompile Include="..\..\Common\ArmCPUDetect.cpp" />
     <ClCompile Include="..\..\Common\ArmEmitter.cpp" />
     <ClCompile Include="..\..\Common\Buffer.cpp" />
+    <ClCompile Include="..\..\Common\Input\GestureDetector.cpp" />
+    <ClCompile Include="..\..\Common\Input\InputState.cpp" />
     <ClCompile Include="..\..\Common\Serialize\Serializer.cpp" />
     <ClCompile Include="..\..\Common\ColorConv.cpp" />
     <ClCompile Include="..\..\Common\ColorConvNEON.cpp" />
@@ -466,6 +475,10 @@
     <ClCompile Include="..\..\Common\OSVersion.cpp" />
     <ClCompile Include="..\..\Common\StringUtils.cpp" />
     <ClCompile Include="..\..\Common\ThreadPools.cpp" />
+    <ClCompile Include="..\..\Common\Thread\Executor.cpp" />
+    <ClCompile Include="..\..\Common\Thread\PrioritizedWorkQueue.cpp" />
+    <ClCompile Include="..\..\Common\Thread\ThreadPool.cpp" />
+    <ClCompile Include="..\..\Common\Thread\ThreadUtil.cpp" />
     <ClCompile Include="..\..\Common\Thunk.cpp" />
     <ClCompile Include="..\..\Common\Timer.cpp" />
     <ClCompile Include="..\..\Common\TimeUtil.cpp" />
