+++ Windows/InputDevice.cpp
@@ -19,7 +19,7 @@
 #include <thread>
 #include <atomic>
 
-#include "input/input_state.h"
+#include "Common/Input/InputState.h"
 #include "Common/Thread/ThreadUtil.h"
 #include "Core/Config.h"
 #include "Core/Host.h"
