+++ Windows/XinputDevice.cpp
@@ -8,8 +8,8 @@
 #include "Common/CommonWindows.h"
 #include "Common/KeyMap.h"
 #include "Common/Log.h"
-#include "input/input_state.h"
-#include "input/keycodes.h"
+#include "Common/Input/InputState.h"
+#include "Common/Input/KeyCodes.h"
 #include "XinputDevice.h"
 #include "Core/Core.h"
 #include "Core/HLE/sceCtrl.h"
