+++ android/jni/app-android.cpp
@@ -53,8 +53,8 @@ struct JNIEnv {};
 #include "base/NativeApp.h"
 #include "Common/Thread/ThreadUtil.h"
 #include "file/zip_read.h"
-#include "input/input_state.h"
-#include "input/keycodes.h"
+#include "Common/Input/InputState.h"
+#include "Common/Input/KeyCodes.h"
 #include "profiler/profiler.h"
 #include "math/math_util.h"
 #include "net/resolve.h"
