+++ ext/native/ui/root.h
@@ -3,7 +3,7 @@
 #include <functional>
 
 #include "ui/ui_context.h"
-#include "input/input_state.h"
+#include "Common/Input/InputState.h"
 
 namespace UI {
