+++ ext/native/ui/screen.h
@@ -17,7 +17,7 @@
 #include <mutex>
 
 #include "base/NativeApp.h"
-#include "input/input_state.h"
+#include "Common/Input/InputState.h"
 
 #include "Common/Common.h"
