+++ ext/native/ui/view.cpp
@@ -1,8 +1,8 @@
 #include <algorithm>
 #include <mutex>
 
-#include "input/input_state.h"
-#include "input/keycodes.h"
+#include "Common/Input/InputState.h"
+#include "Common/Input/KeyCodes.h"
 #include "gfx_es2/draw_buffer.h"
 #include "gfx/texture_atlas.h"
 #include "util/text/utf8.h"
