+++ libretro/Makefile.common
@@ -148,6 +148,8 @@ SOURCES_CXX += \
 	$(COMMONDIR)/Thread/ThreadUtil.cpp \
 	$(COMMONDIR)/Thread/ThreadPool.cpp \
 	$(COMMONDIR)/Thread/PrioritizedWorkQueue.cpp \
+	$(COMMONDIR)/Input/GestureDetector.cpp \
+	$(COMMONDIR)/Input/InputState.cpp \
 	$(COMMONDIR)/Buffer.cpp \
 	$(COMMONDIR)/ConsoleListener.cpp \
 	$(COMMONDIR)/ExceptionHandlerSetup.cpp \
@@ -299,8 +301,6 @@ SOURCES_CXX += $(NATIVEDIR)/math/dataconv.cpp \
 	       $(NATIVEDIR)/i18n/i18n.cpp \
 	       $(NATIVEDIR)/image/zim_load.cpp \
 	       $(NATIVEDIR)/image/png_load.cpp \
-	       $(NATIVEDIR)/input/gesture_detector.cpp \
-	       $(NATIVEDIR)/input/input_state.cpp \
 	       $(NATIVEDIR)/math/curves.cpp \
 	       $(NATIVEDIR)/math/expression_parser.cpp \
 	       $(NATIVEDIR)/math/lin/matrix4x4.cpp \
