+++ libretro/libretro.cpp
@@ -27,7 +27,7 @@
 #include "GPU/GPUInterface.h"
 #include "GPU/Common/FramebufferManagerCommon.h"
 #include "GPU/Common/TextureScalerCommon.h"
-#include "input/input_state.h"
+#include "Common/Input/InputState.h"
 #include "base/NativeApp.h"
 #include "Common/Thread/ThreadUtil.h"
